Make withSocketsDo truly cheap: make it a noop.
This patch implements an alternative way of avoiding explicit calls to
`withSocketsDo` to ensure that the Winsock API on Windows is
initialized, suggested by Simon Marlow in a [blog comment].

The idea is to use linker `constructor`/`destructor` attributes to
initialize Winsock at link/exit time. As a result, `withSocketsDo` is
just the identity function now, retained for backwards compatibility
only. It's not marked deprecated because it's still useful to call it
when building against older versions of network.

This is a followup to #158.

[blog comment]: http://neilmitchell.blogspot.fr/2015/02/making-withsocketsdo-unnecessary.html?showComment=1425048549634#c6282261058772587666